requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.12.0
azure-eventhub==5.10.1
azure-servicebus==7.8.1
opencensus-ext-logging
opencensus-ext-requests
opencensus-extension-azure-functions
protobuf==3.20.*
azure-cosmos==4.3.0
python-dotenv==0.21.1
azure-data-tables==12.4.2
